Release management: add release management bits
We based this on Twisted's own settings, which use hatchling as a
build system.

With these changes, the project is now published as RC on pypi, mostly
to secure the module name.

It has not been tested yet, but we should be able to install using pip
and run the project as a mdoule.